Rewrite Minion ID generator
* Keep order on deduplication
* Fix logical expression
* Rewrite generate_minion_id function
* Use api of the distinct list
* Use FQDN at last, as it may differ
* Rename network unit test into a proper name
* Fix network test for Solaris
* Get hostname as an attribute of a raw socket.
* Make generate_minioin_id more testable.
* Add unit test for check if the minion ID is distinct in the pool
* Add unit test to check on duplicate hostnames in the pool
* Add test for the first in line from platform.node is used (mostly)
* Add unit test to filter out localhost from the first occurrence
* Add unit test to check if any of localhost or local IP addresses are filtered
* Add unit test check for non-localhost IP is accepted as a minion ID
* Fix the documentation
* Return 'localhost' if no other name has been found.
* Check raw socket attribute only if other socket checks had failed.
* Test fix: look for attribute name only when other socket checks had failed
* Test fix: only non-localhost IP should persist
* Add unit test: minion ID should be localhost, if everything is localhost.
* Add unit test: check if FQDN is picked up
* Add unit test: addrinfo from the raw socket should be picked up if all other socket checks had failed to localhost
* Do not use cached localhost name.
